refactor(database): split api routes from services
